requirements.txt
# Production dependencies
# Core packages for your application
openai>=1.93.0
python-dotenv>=1.0.0
pydantic>=2.11.7
httpx>=0.28.1
certifi>=2025.6.15
tqdm>=4.67.1
typing-extensions>=4.14.0
PyMuPDF>=1.23.0